requirements.txt
django==3.2.15
djangorestframework==3.13.1
pillow==9.2.0
numpy==1.23.3